Rewrite the log handler to work asynchronously
Major rewrite of the handler to queue log messages in a Sqlite database
for later asynchronous transmission to Logstash. The actual transmission
happens in a seperate worker thread to pull off the load from the main
or logging thread.
Also split the code in modules and remove unneeded code.

For more details, see README.rst.