Exclude fields early in _get_record_fields
The `_get_record_fields` method extracts and formats the items of the `LogRecord` object into the message dict. After collecting the static extra fields and adding the per-record extra fields, `_remove_excluded_fields` is called to remove any keys we don't want.
`_remove_excluded_fields` is called for the two `dict`s `message` and `extra_fields` and removes items if the key is present in `FORMATTER_RECORD_FIELD_SKIP_LIST`. We can improve this by not adding those fields in the first place since we're iterating over `record.__dict__` anyways.